backend/requirements.txt
# Waveshare E-Paper Library
RPi.GPIO
spidev
Pillow>=10.0.0

# Bluetooth Low Energy
bleak>=0.21.0
dbus-python>=1.3.2

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# Utilities
python-dotenv>=1.0.0
aiofiles>=23.2.0
requests>=2.31.0
pyyaml>=6.0.1

# Image Processing
qrcode>=7.4.2

# Scheduling
APScheduler>=3.10.4
